requirements.txt
SpeechRecognition==3.8.1
moviepy==1.0.3
pydub==0.25.1
tqdm==4.67.1
noisereduce==3.0.2
scipy==1.14.1
numpy==2.0.2
openai-whisper==20231117